+++ CHANGELOG.md
@@ -15,9 +15,12 @@ do not hesitate to take a look at the user manual.
  - Method Joint::getReactionTorque() has been added to retrieve the current reaction torque of a joint
  - Method RigidBody::setLinearLockAxisFactor() to lock the translational movement of a body along the world-space x, y and z axes
  - Method RigidBody::setAngularLockAxisFactor() to lock the rotational movement of a body around the world-space x, y and z axes
- - Method RigidBody::applyLocalForceAtWorldPosition()
- - Method RigidBody::applyLocalForceAtLocalPosition()
- - Method RigidBody::applyLocalForceToCenterOfMass()
+ - Method RigidBody::applyLocalForceAtWorldPosition() to manually apply a force to a rigid body
+ - Method RigidBody::applyLocalForceAtLocalPosition() to manually apply a force to a rigid body
+ - Method RigidBody::applyLocalForceToCenterOfMass() to manually apply a force to a rigid body
+ - Method RigidBody::applyLocalTorque() to apply a local-space torque to a rigid body
+ - Method RigidBody::getForce() to get the total manually applied force on a rigid body
+ - Method RigidBody::getTorque() to get the total manually applied torque on a rigid body
  - A cone limit can now be set to the ball-and-socket joint (this is useful for ragdolls)
  - Bridge scene has been added to the testbed application
  - Ragdoll scene has been added to the testbed application
@@ -32,6 +35,7 @@ do not hesitate to take a look at the user manual.
  - Rename method RigidBody::applyForceAtWorldPosition() into RigidBody::applyWorldForceAtWorldPosition()
  - Rename method RigidBody::applyForceAtLocalPosition() into RigidBody::applyWorldForceAtLocalPosition()
  - Rename method RigidBody::applyForceToCenterOfMass() into RigidBody::applyWorldForceAtCenterOfMass()
+ - Rename method RigidBody::applyTorque() into RigidBody::applyWorldTorque()
  - The raycasting broad-phase performance has been improved
  - The raycasting performance against HeighFieldShape has been improved (better middle-phase algorithm)
  - Robustness of polyhedron vs polyhedron collision detection has been improved in SAT algorithm (face contacts are favored over edge-edge contacts for better stability)